+++ sdk/storage/azure_storage_blob/src/clients/block_blob_client.rs
@@ -74,6 +74,7 @@ impl BlockBlobClient {
         options: Option<BlockBlobClientOptions>,
     ) -> Result<Self> {
         let mut options = options.unwrap_or_default();
+        super::apply_client_defaults(&mut options.client_options);
         apply_storage_logging_defaults(&mut options.client_options);
 
         let storage_headers_policy = Arc::new(StorageHeadersPolicy);