+++ sdk/storage/azure_storage_blob/tests/blob_service_client.rs
@@ -217,7 +217,7 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
     .await?;
 
     // Sleep in live mode to allow tags to be indexed on the service
-    if recording.test_mode() == TestMode::Live {
+    if recording.test_mode() != TestMode::Playback {
         time::sleep(Duration::from_secs(5)).await;
     }